chore: renamed initRN prop, wagmiConfig file, updated web-components versions, added types for wagmiConfig, removed balance from connect button
