+++ package.json
@@ -12,6 +12,8 @@
     "@next/third-parties": "^14.2.5",
     "@radix-ui/react-dialog": "^1.1.0",
     "@rainbow-me/rainbowkit": "^2.2.0",
+    "@requestnetwork/create-invoice-form": "^0.9.0",
+    "@requestnetwork/invoice-dashboard": "^0.8.0",
     "@requestnetwork/payment-detection": "0.43.1-next.2043",
     "@requestnetwork/request-client.js": "0.47.1-next.2043",
     "@requestnetwork/web3-signature": "0.6.1-next.2043",
