+++ pages/_app.tsx
@@ -6,7 +6,7 @@ import { GoogleTagManager } from "@next/third-parties/google";
 import { QueryClientProvider, QueryClient } from "@tanstack/react-query";
 import { Provider } from "@/utils/context";
 import Navbar from "@/components/common/Navbar";
-import { rainbowKitConfig } from "@/utils/connectWallet";
+import { rainbowKitConfig } from "@/utils/wagmiConfig";
 import VersionDisplay from "@/components/common/VersionBadge";
 import "@/styles/globals.css";
 import "@rainbow-me/rainbowkit/styles.css";
