+++ pages/create-invoice.tsx
@@ -1,9 +1,9 @@
 import Head from "next/head";
-import CreateInvoiceForm from "@requestnetwork/create-invoice-form/react";
 import { config } from "@/utils/config";
 import { useAppContext } from "@/utils/context";
 import { currencies } from "@/utils/currencies";
-import { rainbowKitConfig as wagmiConfig } from "@/utils/connectWallet";
+import { rainbowKitConfig as wagmiConfig } from "@/utils/wagmiConfig";
+import CreateInvoiceForm from "@requestnetwork/create-invoice-form/react";
 
 export default function CreateInvoice() {
   const { requestNetwork } = useAppContext();
